+++ tests/test_models.py
@@ -1,14 +1,39 @@
 # -*- coding: utf-8 -*-
+from copy import deepcopy
 from unittest import mock
 
+from django.conf import settings
+from django.core.cache import caches
 from django.core.exceptions import ValidationError
-from django.test import TestCase
+from django.test import TestCase, override_settings
 
-from cookie_consent.cache import delete_cache
+from cookie_consent.cache import CACHE_KEY, delete_cache
 from cookie_consent.models import Cookie, CookieGroup, validate_cookie_name
 
+patch_caches = override_settings(
+    COOKIE_CONSENT_CACHE_BACKEND="tests",
+    CACHES={
+        **deepcopy(settings.CACHES),
+        "tests": {
+            "BACKEND": "django.core.cache.backends.locmem.LocMemCache",
+        },
+    },
+)
 
-class CookieGroupTest(TestCase):
+
+class CacheMixin:
+    def populateCache(self):
+        cache = caches["tests"]
+        cache.set(CACHE_KEY, {}, timeout=3600)
+
+    def assertCacheNotPopulated(self):
+        cache = caches["tests"]
+        has_key = cache.has_key(CACHE_KEY)
+        self.assertFalse(has_key)
+
+
+@patch_caches
+class CookieGroupTest(CacheMixin, TestCase):
     def setUp(self):
         self.addCleanup(delete_cache)
 
@@ -26,27 +51,30 @@ class CookieGroupTest(TestCase):
             self.cookie_group.get_version(), self.cookie.created.isoformat()
         )
 
-    @mock.patch("cookie_consent.models.delete_cache")
-    def test_bulk_delete(self, mock_delete_cache):
+    def test_bulk_delete(self):
+        self.populateCache()
+
         deleted_objs_count, _ = CookieGroup.objects.filter(
             id=self.cookie_group.id
         ).delete()
 
         # Deleting a CookieGroup also deletes the associated Cookies, that's why we expect a count of 2.
         self.assertEqual(deleted_objs_count, 2)
-        self.assertEqual(mock_delete_cache.call_count, 1)
+        self.assertCacheNotPopulated()
 
-    @mock.patch("cookie_consent.models.delete_cache")
-    def test_bulk_update(self, mock_delete_cache):
+    def test_bulk_update(self):
+        self.populateCache()
+
         updated_objs_count = CookieGroup.objects.filter(id=self.cookie_group.id).update(
             name="Optional2"
         )
 
         self.assertEqual(updated_objs_count, 1)
-        self.assertEqual(mock_delete_cache.call_count, 1)
+        self.assertCacheNotPopulated()
 
 
-class CookieTest(TestCase):
+@patch_caches
+class CookieTest(CacheMixin, TestCase):
     def setUp(self):
         self.addCleanup(delete_cache)
 
@@ -63,21 +91,23 @@ class CookieTest(TestCase):
     def test_varname(self):
         self.assertEqual(self.cookie.varname, "optional=foo:.example.com")
 
-    @mock.patch("cookie_consent.models.delete_cache")
-    def test_bulk_delete(self, mock_delete_cache):
+    def test_bulk_delete(self):
+        self.populateCache()
+
         deleted_objs_count, _ = Cookie.objects.filter(id=self.cookie.id).delete()
 
         self.assertEqual(deleted_objs_count, 1)
-        self.assertEqual(mock_delete_cache.call_count, 1)
+        self.assertCacheNotPopulated()
 
-    @mock.patch("cookie_consent.models.delete_cache")
-    def test_bulk_update(self, mock_delete_cache):
+    def test_bulk_update(self):
+        self.populateCache()
+
         updated_objs_count = Cookie.objects.filter(id=self.cookie.id).update(
             name="foo2"
         )
 
         self.assertEqual(updated_objs_count, 1)
-        self.assertEqual(mock_delete_cache.call_count, 1)
+        self.assertCacheNotPopulated()
 
 
 class ValidateCookieNameTest(TestCase):